
\name{Datos_Prueba}
\alias{Datos_Prueba}
\title{Datos_Prueba (Example dataset)}

\description{
Dataset of foods and their estimated retail prices per 100g for Cali, Colombia, in September 2022. In addition to the foods, it includes the energy content and various nutrient information
}

\usage{
Datos_Prueba
}

\format{

Dataset with 99 foods (rows) and 24 variables (columns). The first columns provide information about the food, including "Cod_TCAC" (Food code according to TCAC), exchanges, price per exchange, and the energy and nutrient contents for each

}

\examples{

library(Foodprice)
head(Datos_Prueba)

}

\source{


}