
\name{diverse}
\alias{diverse}
\title{diverse (Example dataset)}

\description{
Dataset containing the quantity of foods to be selected per groups according to GABAS.
\usage{
diverse
}

\format{
Dataset with two columns (Group and Quantity) and ten rows (food groups).
}

\examples{

library(Foodprice)
head(diverse)

}

\source{
- GABAS: Instituto Colombiano de Bienestar Familiar. (2020). Guías Alimentarias Basadas en Alimentos para la población colombiana mayor de 2 años. En ICBF/FAO (2da Edic.). Recovered from https://www.minsalud.gov.co/sites/rid/Lists/BibliotecaDigital/RIDE/VS/PP/SNA/guiasalimentariasbasadas-en-alimentos.pdf

}