
\name{EER_example}
\alias{EER_example}
\title{(Example dataset)-EER_example}

\description{
Sample dataset with energy requirements for arbitrary demographic groups and genders. The values are the same as those established for all of Colombia in the 'EER' data.
}

\usage{
EER_example
}

\examples{

library(Foodprice)
head(EER_example)

}

\source{


}

