
\name{EER_LL_example}
\alias{EER_LL_example}
\title{(Example dataset)-EER_LL_example}

\description{

Sample dataset with energy requirements and minimum requirements for various nutrients in arbitrary demographic groups and genders. The requirements are the same as those established for all of Colombia in the 'EER_LL' data.}

\usage{
EER_LL_example
}

\examples{

library(Foodprice)
head(EER_LL_example)

}

\source{


}

