
\name{exclusion_3er_modelo}
\alias{exclusion_3er_modelo}
\title{exclusion_3er_modelo (dataset for DataCol function)}

\description{
Dataset for filtering foods with very low caloric content, condiments consumed in small quantities, herbs, spices, or preparations.
}

\usage{
exclusion_3er_modelo
}

\format{
Dataset of 4 foods (rows) and two columns, "Food" (Food) and "Cod_TCAC" (TCAC Code).
}

\examples{

library(Foodprice)
head(exclusion_3er_modelo)

}


