
\name{UL_example}
\alias{UL_example}
\title{(Example dataset)-UL_example}

\description{

Dataset with energy requirements and maximum requirements for various nutrients in different demographic groups and arbitrary genders. The energy requirement values are the same as those in the 'UL' dataset.
}
\usage{
UL_example
}


\examples{

library(Foodprice)
head(UL_example)

}

\source{


}