#------------------------------------------------------------------------------------------#
#                 D O C U M E N T A C I Ó N: FUNCIÓN-Datos_dane_col                        #
#-----------------------------------------------------------------------------------------#


\name{Datos_dane_col}
\alias{Datos_dane_col}
\title{Datos_dane_col}
\usage{
Datos_dane_col()
}
\description{
Prints 'Datos_dane_col, world!'.
}
\examples{
Datos_dane_col()
}
