
\name{serv_example}
\alias{serv_example}
\title{(Dataset of requirements for Colombia)-serv_example}

\description{
Dataset containing exchange requirements for each food subgroup, demographic group, and gender (arbitrarily created). The exchange values are the same as those in the 'serv' dataset.

\usage{
serv_example
}


\examples{

library(Foodprice)
head(serv_example)

}

\source{


}